+++ ErsatzTV.Application/Playouts/PlayoutNameViewModel.cs
@@ -1,8 +1,12 @@
-﻿namespace ErsatzTV.Application.Playouts
+﻿using System;
+using LanguageExt;
+
+namespace ErsatzTV.Application.Playouts
 {
     public record PlayoutNameViewModel(
         int PlayoutId,
         string ChannelName,
         string ChannelNumber,
-        string ScheduleName);
+        string ScheduleName,
+        Option<TimeSpan> DailyRebuildTime);
 }